CO2data

Compiled records of atmospheric CO2 concentrations and stable carbon isotopes to reconstruct climate and derive plant ecophysiological indices from tree rings. Data are from Belmecheri, Lavergne, 2020, Dendrochronologia. Updated based on their methodology beyond C.E. 2019.

data(CO2data)

Format

A data frame with 2023 rows and 3 variables:

yr

Year of CO2 and d13CO2 measurement

Ca

Atmospheric CO2 concentration, in ppm

d13C.atm

Atmospheric d13CO2, in per mille, ‰

Source

https://www.sciencedirect.com/science/article/abs/pii/S1125786520300874

References

Belmecheri, S. & Lavergne, A. (2020). Compiled records of atmospheric CO2 concentrations and stable carbon isotopes to reconstruct climate and derive plant ecophysiological indices from tree rings. Dendrochronologia, 63, 125748.

Examples

data(CO2data)
head(CO2data)
#>   yr     Ca d13C.atm
#> 1  0 277.63    -6.41
#> 2  1 277.63    -6.41
#> 3  2 277.64    -6.41
#> 4  3 277.64    -6.41
#> 5  4 277.65    -6.41
#> 6  5 277.66    -6.41
tail(CO2data)
#>        yr       Ca  d13C.atm
#> 2018 2017 404.3096 -8.538528
#> 2019 2018 406.1517 -8.536936
#> 2020 2019 408.8255 -8.579904
#> 2021 2020 411.0815 -8.591044
#> 2022 2021 413.0753 -8.594226
#> 2023 2022 415.0163 -8.656291